+++ lib/modal_progress_indicator.dart
@@ -1,3 +1,6 @@
+/// Configurable modal progress indicator that blocks access to
+/// a wrapped widget and its subtree.
 library modal_progress_indicator;
 
+export 'src/background_options.dart';
 export 'src/modal_progress_indicator.dart';