+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.0.0
 
-- A modal progress indicator that blocks access to widgets while it's visible.
+- Configurable modal progress indicator that blocks access to a wrapped widget and its subtree.